+++ apps/sb-vite/.storybook/preview.tsx
@@ -5,15 +5,7 @@ import 'raf/polyfill' // const OriginalNextImage = NextImage.default
 import { Preview } from '@storybook/react'
 import { StorybookDecorator } from 'app/src/provider/StorybookDecorator'
 import i18n from './i18n'
-import { RouterContext } from 'next/dist/shared/lib/router-context.shared-runtime'
 
-const RouterDecorator = (Story) => {
-  return (
-    <RouterContext.Provider value={RouterContext}>
-      <Story />
-    </RouterContext.Provider>
-  )
-}
 const preview: Preview = {
   globalTypes: {
     theme: {
@@ -45,7 +37,7 @@ const preview: Preview = {
     //   // Provider: RouterContext.Provider, // next 13 (using next/router) / next < 12
     // },
   },
-  decorators: [RouterDecorator, StorybookDecorator],
+  decorators: [StorybookDecorator],
 }
 
 export default preview