+++ packages/form/src/LmFormFieldContainer.tsx
@@ -47,15 +47,15 @@ export function LmFormFieldContainer({
   ...rest
 }: LmFormContainerProps) {
   return (
-    <StackContainer space={rest.labelInline ? '$3' : undefined} {...rest}>
+    <StackContainer {...rest} space={rest.labelInline ? '$3' : rest.space}>
       {label && (
         <Label
           htmlFor={id}
           size={size || '$3'}
           {...labelProps}
-          color={error ? '$red10' : undefined}
-          width={rest.labelInline ? 150 : undefined}
-          justifyContent={rest.labelInline ? 'flex-end' : undefined}
+          color={error ? '$red10' : labelProps?.color}
+          width={rest.labelInline ? 150 : labelProps?.width}
+          justifyContent={rest.labelInline ? 'flex-end' : labelProps?.justifyContent}
         >
           {label} {required && ` *`}
         </Label>